requirements.txt
Django==1.8.6
django-apscheduler==0.3.0
pyOpenSSL==18.0.0
django-sslserver==0.20
line-bot-sdk==1.12.1
zeep==3.4.0

python-barcode==0.10.0
imgurpython==1.1.7